pyproject.toml
[project]
name = "admissions-conversation-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "haroldvillota", email = "aroldf@gmail.com" }
]
requires-python = ">=3.11,<3.12"
dependencies = [
    "langchain>=1.2.7",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-postgres>=0.0.16",
    "langfuse>=3.12.1",
    "langgraph>=1.0.7",
    "langgraph-api>=0.7.22",
    "langgraph-checkpoint>=4.0.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langgraph-cli[inmem]>=0.4.12",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "typing-extensions>=4.15.0",
]

[project.scripts]
cli = "admissions_conversation_engine.entrypoints.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
